migrate load /write classes from unorganized to py
